Placing query in external file to make it easier to find
