+++ packages/financial-templates-lib/src/price-feed/BigQueryPriceFeed.js
@@ -3,6 +3,7 @@ const { BigQuery } = require("@google-cloud/bigquery");
 const moment = require("moment");
 const highland = require("highland");
 const { parseFixed } = require("@ethersproject/bignumber");
+const { createQuery } = require("../queries/GasEthQuery");
 
 const client = new BigQuery();
 
@@ -55,7 +56,7 @@ class BigQueryPriceFeed extends PriceFeedInterface {
       .format("YYYY-MM-DD HH:mm:ss");
 
     // Create the query with the needed time interval.
-    const query = this.createQuery(t2, formattedCurrentTime);
+    const query = createQuery(t2, formattedCurrentTime);
 
     // Submit async call to BigQuery and check the response.
     let priceResponse;
@@ -112,7 +113,7 @@ class BigQueryPriceFeed extends PriceFeedInterface {
     });
 
     // Create the query with the needed time interval.
-    const query = this.createQuery(t2, formattedCurrentTime);
+    const query = createQuery(t2, formattedCurrentTime);
 
     // Submit async call to BigQuery and check response.
     let priceResponse;
@@ -145,64 +146,6 @@ class BigQueryPriceFeed extends PriceFeedInterface {
         // this could also "pipe" into some other processing pipeline or write to a file
         .toPromise(Promise)
     );
-  }
-  // This is a helper method to create a GASETH BQ query with time arguments.
-  createQuery(t2, formattedCurrentTime) {
-    const query = `
-        DECLARE halfway int64;
-        DECLARE block_count int64;
-        DECLARE max_block int64;
-
-        -- Querying for the amount of blocks in the preset time range. This will allow block_count to be compared against a given minimum block amount.
-        SET (block_count, max_block) = (SELECT AS STRUCT (MAX(number) - MIN(number)), MAX(number) FROM \`bigquery-public-data.crypto_ethereum.blocks\` 
-        WHERE timestamp BETWEEN TIMESTAMP('${t2}', 'UTC') AND TIMESTAMP('${formattedCurrentTime}', 'UTC'));
-
-        CREATE TEMP TABLE cum_gas (
-        gas_price int64,
-        cum_sum int64
-        );
-
-        -- If the minimum threshold of blocks is met, query on a time range
-        IF block_count >= 134400 THEN
-        INSERT INTO cum_gas (
-        SELECT
-            gas_price,
-            SUM(gas_used) OVER (ORDER BY gas_price) AS cum_sum
-        FROM (
-            SELECT
-            gas_price,
-            SUM(receipt_gas_used) AS gas_used
-            FROM
-            \`bigquery-public-data.crypto_ethereum.transactions\`
-            WHERE block_timestamp 
-            BETWEEN TIMESTAMP('${t2}', 'UTC')
-            AND TIMESTAMP('${formattedCurrentTime}', 'UTC')  
-            GROUP BY
-            gas_price));
-        ELSE -- If a minimum threshold of blocks is not met, query for the minimum amount of blocks
-        INSERT INTO cum_gas (
-        SELECT
-            gas_price,
-            SUM(gas_used) OVER (ORDER BY gas_price) AS cum_sum
-        FROM (
-            SELECT
-            gas_price,
-            SUM(receipt_gas_used) AS gas_used
-            FROM
-            \`bigquery-public-data.crypto_ethereum.transactions\`
-            WHERE block_number 
-            BETWEEN (max_block - 134400)
-            AND max_block
-            GROUP BY
-            gas_price));
-        END IF;
-
-        SET halfway = (SELECT DIV(MAX(cum_sum),2) FROM cum_gas);
-
-        SELECT cum_sum, gas_price FROM cum_gas WHERE cum_sum > halfway ORDER BY gas_price LIMIT 1;
-    `;
-
-    return query;
   }
 }
